# Add defer_build to pydantic models for faster aiida.orm import time