@danielhollas 2025-05-07
Just a few more benchmarks, comparing this PR versus AiiDA 2.6.3 as an observation:

### verdi profile list

This one is a bit suspicious:

```
❯ hyperfine -w 1 "uv run --no-config --directory /tmp verdi profile list" "uv run verdi profile list"
Benchmark 1: uv run --no-config --directory /tmp verdi profile list
  Time (mean ± σ):     192.3 ms ±   2.8 ms    [User: 163.4 ms, System: 26.7 ms]
  Range (min … max):   187.8 ms … 195.8 ms    15 runs
 
Benchmark 2: uv run verdi profile list
  Time (mean ± σ):     361.3 ms ±   3.8 ms    [User: 305.5 ms, System: 52.1 ms]
  Range (min … max):   357.0 ms … 368.9 ms    10 runs
```

### verdi computer list

```
❯ hyperfine -w 1 "uv run --no-config --directory /tmp verdi computer list" "uv run verdi computer list"
Benchmark 1: uv run --no-config --directory /tmp verdi computer list
  Time (mean ± σ):     748.1 ms ±   4.9 ms    [User: 736.4 ms, System: 838.4 ms]
  Range (min … max):   742.4 ms … 756.8 ms    10 runs
 
Benchmark 2: uv run verdi computer list
  Time (mean ± σ):     977.6 ms ±   6.5 ms    [User: 937.5 ms, System: 860.2 ms]
  Range (min … max):   968.2 ms … 991.2 ms    10 runs
 
Summary
  uv run --no-config --directory /tmp verdi computer list ran
    1.31 ± 0.01 times faster than uv run verdi computer list
```