@khsrali 2025-05-08
Thanks @danielhollas, very interesting!

About your latest comment (`verdi computer list & verdi profile list`) 
Do you also have the numbers for the current main? I assume they're terribly slow(?)