@danielhollas 2025-05-08
They are about 100ms slower than this PR.