@danielhollas 2025-05-08
Note that I also tried adding `defer_build` to other orm classes but didn't see significant speedup. 
I think it's because the `Sealable` and `Entity` classes are mixin/abstract classes that are used in many other concrete Node classes so the respective pydantic `Model` class gets build many times.

I tried to analyze a bit more the cProfile data to see where the rest of the 100ms time is spent, but didn't see anything obvious.