@edan-bainglass 2025-05-09
> Part of the slow down (20ms) is that we're now calling the `hasattr` method much more (2k more calls) (using cProfile and pstats modules)

I wonder how much of this is coming from the construction of `fields` in `EntityFieldMeta`. If so, maybe this part could be improved? Perhaps regardless of the burden...