@danielhollas 2025-05-09
Yes, I had actually looked into EntityFieldsMeta, but I didn't see an
obvious way to improve it. And switching from hasattr to getattr didn't
improve things.

On Fri, May 9, 2025, 07:09 Edan Bainglass ***@***.***> wrote:

> *edan-bainglass* left a comment (aiidateam/aiida-core#6867)
> <https://github.com/aiidateam/aiida-core/pull/6867#issuecomment-2865253374>
>
> Part of the slow down (20ms) is that we're now calling the hasattr method
> much more (2k more calls) (using cProfile and pstats modules)
>
> I wonder how much of this is coming from the construction of fields in
> EntityFieldMeta. If so, maybe this part could be improved? Perhaps
> regardless of the burden...
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/aiidateam/aiida-core/pull/6867#issuecomment-2865253374>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ACIY64KNHDHXCWM3M7G4WHT25RBBNAVCNFSM6AAAAAB4UWUGA2VHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMZDQNRVGI2TGMZXGQ>
> .
> You are receiving this because you were mentioned.Message ID:
> ***@***.***>
>