@danielhollas 2025-05-12
I am no longer able to reproduce any difference for `verdi profile list`

```
❯ hyperfine -w 5 "uv run --no-config --directory /tmp/2.6 verdi profile list" "uv run --no-config --no-sync verdi profile list"
Benchmark 1: uv run --no-config --directory /tmp/2.6 verdi profile list
  Time (mean ± σ):     193.9 ms ±   4.2 ms    [User: 163.0 ms, System: 28.9 ms]
  Range (min … max):   187.1 ms … 199.6 ms    15 runs
 
Benchmark 2: uv run --no-config --no-sync verdi profile list
  Time (mean ± σ):     195.5 ms ±   4.0 ms    [User: 163.6 ms, System: 30.0 ms]
  Range (min … max):   189.8 ms … 204.3 ms    15 runs
 
Summary
  uv run --no-config --directory /tmp/2.6 verdi profile list ran
    1.01 ± 0.03 times faster than uv run --no-config --no-sync verdi profile list
```

And the difference I am seeing now for `aiida.orm` import between this branch and 2.6.3 is more like `50ms` so I don't think it's worth chasing further at this point. 

```console
❯ hyperfine -w 5 "uv run --no-config --directory /tmp/2.6 python -c 'import aiida.orm'" "uv run --no-config --no-sync python -c 'import aiida.orm'"
Benchmark 1: uv run --no-config --directory /tmp/2.6 python -c 'import aiida.orm'
  Time (mean ± σ):     265.6 ms ±   6.5 ms    [User: 296.1 ms, System: 797.5 ms]
  Range (min … max):   260.5 ms … 283.9 ms    11 runs
 
Benchmark 2: uv run --no-config --no-sync python -c 'import aiida.orm'
  Time (mean ± σ):     319.5 ms ±   2.5 ms    [User: 348.5 ms, System: 803.3 ms]
  Range (min … max):   317.0 ms … 323.3 ms    10 runs
 
Summary
  uv run --no-config --directory /tmp/2.6 python -c 'import aiida.orm' ran
    1.20 ± 0.03 times faster than uv run --no-config --no-sync python -c 'import aiida.orm'
```

(it's possible that some of the overhead I was seeing before came from `uv run` itself.)