@danielhollas 2025-05-15
Yes. This meaningfully speeds up the import times.