@edan-bainglass 2025-11-08
@danielhollas I'm working on the Pydantic system and wanted to rerun some benchmarks following your work here. However, I get the following:
```
(aiida-dev) (main) aiida-core > hyperfine -w 1 "uv run --no-config --directory /tmp python -c 'import aiida.orm'" "uv run python -c 'import aiida.orm'" --show-output
Benchmark 1: uv run --no-config --directory /tmp python -c 'import aiida.orm'
  Time (mean ± σ):     482.3 ms ±  22.2 ms    [User: 960.0 ms, System: 1426.6 ms]
  Range (min … max):   459.7 ms … 537.6 ms    10 runs
 
Benchmark 2: uv run python -c 'import aiida.orm'
Traceback (most recent call last):
  File "<string>", line 1, in <module>
  File "/home/edanb/AiiDA/aiida-core/src/aiida/orm/__init__.py", line 23, in <module>
    from .nodes import *
  File "/home/edanb/AiiDA/aiida-core/src/aiida/orm/nodes/__init__.py", line 16, in <module>
    from .data import *
  File "/home/edanb/AiiDA/aiida-core/src/aiida/orm/nodes/data/__init__.py", line 30, in <module>
    from .remote import *
  File "/home/edanb/AiiDA/aiida-core/src/aiida/orm/nodes/data/remote/__init__.py", line 7, in <module>
    from .base import *
  File "/home/edanb/AiiDA/aiida-core/src/aiida/orm/nodes/data/remote/base.py", line 20, in <module>
    from aiida.transports import Transport
  File "/home/edanb/AiiDA/aiida-core/src/aiida/transports/__init__.py", line 15, in <module>
    from .plugins import *
  File "/home/edanb/AiiDA/aiida-core/src/aiida/transports/plugins/__init__.py", line 15, in <module>
    from .async_backend import *
  File "/home/edanb/AiiDA/aiida-core/src/aiida/transports/plugins/async_backend.py", line 24, in <module>
    import asyncssh
  File "/home/edanb/AiiDA/aiida-core/.venv/lib/python3.10/site-packages/asyncssh/__init__.py", line 36, in <module>
    from .channel import SSHClientChannel, SSHServerChannel
ValueError: source code string cannot contain null bytes
Error: Command terminated with non-zero exit code 1 in the first warmup run. Use the '-i'/'--ignore-failure' option if you want to ignore this. Alternatively, use the '--show-output' option to debug what went wrong.
```

Two things:
1. Benchmark 1 is roughly twice as slow as in this PR. However, it appears to be unrelated to my work, as it is the same on `main`
2. Any ideas regarding the error in benchmark 2?